add grenade aim settings - to be tested
RandomizerSettings and RandomizerRebinding can be deleted and will be
re-created with default settings.

In RandomizerSettings, "Instant Grenade Aim: True" will cause grenades
to be aimed at the position of the left analog stick - neutral will
cause a weak throw straight up, tilting the stick fully will aim at the
edges of the screen.
"Grenade Aim Speed: <float>" will multiply the speed at which your aim
adjusts with the normal controller grenade aim when instant aiming is
false.

In RandomizerRebinding, a new bind has been added to reset grenade aim
to the default (slightly up and forward). This does not have a default
binding.